refactor: Clarify file name's role
