+++ src/stack.rs
